SoundCusesDiegetic
Diegetic and Non-diegetic, 5 sounds + a text file describing attributes for the 3 of them under crreative commons.

Though 267250__berlingamescene__berlingamescene(TheOriginal_it_has_a_fadeOut)
Sound cue Diegetic cannot be looped at the moment, proved harder to find a good place to cut it might have to redo that one later. But it is suposed to  play once as a cue it works find but for blending into the same crowd noise I might have to find something else to replace it with. A fix would be the crowd having multiple audiosources to mix between them starting points to ensure the fade cannot be heard.